front-webapp/requirements.txt
flask
requests
azure-identity